Mbnavadjust: Added integrated mbgrdviz-style visualization of the bathymetry in an Mbnavadjust project. Users can select crossings or ties by clicking on the navigation in the map view. If two selected sections do not already have a defined crossing, a crossing can be created and analyzed. The new version of mbnavadjust will upgrade old projects to include the bathymetry grid. The new GUI commands include <Update Grids> under the <Action> menu, and <Visualize Survey> under the <View> menu. The documentation, such as it is, does not yet cover these new features.
Mbm_multicopy and mbm_multidatalist: Fixed bugs identified (along with fixes) by
Martin Ernst of HafenCity University in Hamburg, Germany.

Mbswath, mbcontour, mbgrdtiff: added #ifdefs for compatibility with the upcoming
GMT 5.3 release. A number of GMT functions called by the MB-System GMT modules
are being renamed.

Format 21 (MBF_HSATLRAW): Added some logic to handle problems in a very old HSDS
dataset collected on the R/V Ewing and available at the archive formally known
as NGDC.

+++ src/mbeditviz/mbeditviz_callbacks.c
@@ -1316,13 +1316,13 @@ fprintf(stderr,"do_mbeditviz_viewgrid\n");
 				if ((navstbdlat = malloc(file->num_pings*sizeof(double))) == NULL)
 					mbev_status = MB_FAILURE;
 				if (mbev_status == MB_SUCCESS)
-				if ((navline = malloc(file->num_pings*sizeof(double))) == NULL)
+				if ((navline = malloc(file->num_pings*sizeof(int))) == NULL)
 					mbev_status = MB_FAILURE;
 				if (mbev_status == MB_SUCCESS)
-				if ((navshot = malloc(file->num_pings*sizeof(double))) == NULL)
+				if ((navshot = malloc(file->num_pings*sizeof(int))) == NULL)
 					mbev_status = MB_FAILURE;
 				if (mbev_status == MB_SUCCESS)
-				if ((navcdp = malloc(file->num_pings*sizeof(double))) == NULL)
+				if ((navcdp = malloc(file->num_pings*sizeof(int))) == NULL)
 					mbev_status = MB_FAILURE;
 
 				/* copy nav data */
